Don't URL-escape branch directory names
Previously, if a branch contained (say) an apostrophe or a percent sign,
the directory for that branch would be URL-escaped, with the apostrophe
escaped as %27 or the percent sign escaped as %25.

However, when the web server receives a request, it URL-decodes the
branch name before looking the path up on disk. So the links pointed to
the wrong path. (And I was not able to fetch the correct path in my
browser by requesting %2527 - the browser converts this to %27 before
making the request.)

By definition, Git branch names are valid Unix file paths: that's how
they're stored in the repository. So they don't need to be URL-escaped.

Remove the escaping. Use Jinja2's built-in urlescape filter in the
template. Add a trailing slash to the URL; without it, GitHub Pages
responds to (e.g.):

    GET https://example.github.io/example/%25-blahblah

with 301 Moved Permanently to:

    Location:  https://example.github.io/example/%-blahblah/

which it then serves a 400 Bad Request from because %- is not a valid
URL-encoded character. Thanks, GitHub!

Fixes https://github.com/endlessm/amalgamate-pages/issues/7