Ignore artifacts for deleted branches
Previously, the latest extant artifact for any branch name would be
included in the output, even if that branch has subsequently been
deleted.

This left a lot of clutter in the resulting index, and made it hard to
remove a branch from the output. (You could do so by deleting all
artifacts for that branch but that is tedious.)

Instead, list which branches currently exist in each fork of the
repository, and ignore any artifacts for branches that do not exist.