Delay fetching artifact details until needed
Previously, for each workflow run, the artifact details (though not the
contents!) were fetched eagerly, before checking whether the branch
still exists or whether we already have a newer artifact for this
branch.

Defer fetching artifact details until we know the branch exists and we
don't already have an artifact for it. This dramatically reduces the
number of API requests made, from one per workflow run to one per
branch, and should speed up the generation process as a result.