Replace duplicate files with symlinks
When publishing multiple branches of the same project, it is likely that
some files will be identical between branches. For example, if the
project is a Godot game being published for the web, then in almost
every case the index.wasm file containing the WASM build of the Godot
runtime will be identical in all branches. (The exception would be when
a branch switches to a different Godot version.)

GitHub Pages supports symbolic links. So, we can reduce the size of the
artifact we publish to GitHub Pages by scanning the tree for duplicate
files and replacing them with symbolic links to the first instance of
each file encountered.

This will have no effect on the resulting website, and it could be that
GitHub Pages internally performs some similar duplicate-content
optimization, but it means the resulting artifact will be smaller
(consuming less of the project's artifact quota) and will be uploaded
more quickly (which might reduce the time between pushing a branch and
seeing it published, particularly in projects with many branches).