Show pull request for each branch, if any
The branch name may or may not be particularly useful by itself. If a
branch corresponds to a pull request, show that pull request's title and
link to the pull request's page.

In edge cases there may be multiple pull requests associated with a
branch. In that case, we prefer an open PR to a closed PR, and a more
recently updated PR to an older one.

The index itself remains stubbornly unstyled.