Ignore deleted forks
In order to ignore builds for branches that have been deleted, we
enumerate all live branches on forks that we discover when looking
through builds. However, if a build previously ran for a fork that has
since been deleted, the attempt to list that fork's branches fails with
404 Client Error: Not Found.

Handle this case, treating it the same as if there were no live branches
on that fork (which is true!).

Fixes https://github.com/endlessm/amalgamate-pages/issues/25