Ignore builds where the most recent PR is closed
If a pull request is closed, it follows that the change was either
accepted (in which case it is now in main) or rejected.

In our own projects, we try to delete branches once the corresponding PR
is closed but contributors may not do this. Treat branches whose most
recent PR is closed as if they were deleted.

Fixes https://github.com/endlessm/amalgamate-pages/issues/21