Suppress never-built branches from forks
Typically (always?), the only reason why a branch from a fork would be
built in the upstream is for a pull request. But as soon as we discover
a build from one such third-party branch, we look up all live branches
in that fork. As a result we would list all live branches in all forks,
with "(never built)".

Ignore these.