Reduce special-casing of default branch
Previously, the default branch (typically `main`) of the repository
where the action is running was placed at the root of the site, while
all other pages were placed in the `branches/` directory.

Instead, treat `main` mostly like other branches. Place it in the
`branches/` directory, and list it in the index. But ensure that
branches in the upstream repo are listed before those from forks; and
that the default branch is listed before other branches; so that the
default branch is at the top of the list.

At the root of the site, place a redirect (via `<meta>` tag — GitHub
Pages does not support real redirects) to the `branches/` page. A future
change will put the latest release build, if one exists, at the top
level instead.

In passing, this handles the case where all builds for the default
branch have expired. Previously, the script would crash. Now, the
default branch is shown as expired in the index, like any other branch.

Fixes https://github.com/endlessm/amalgamate-pages/issues/14