Reduce special-casing of default branch
Previously, the default branch (typically `main`) of the repository
where the action is running was placed at the root of the site, and not
listed in the `branches/` directory's index. As a side-effect of how
this was implemented, the script would crash if the default branch had
never been built or if all build artifacts for it had expired.

Instead, treat `main` more like other branches. Unpack it in the same
loop as all branches, taking care to list branches in the upstream repo
before those from forks, and the default branch before other branches,
so that it is at the top of the list.

Place it at the root of the site, if found. Otherwise, redirect from the
root of the site to the `branches/` page (via `<meta>` tag — GitHub
Pages does not support real redirects).

A future change will put the latest release build, if one exists, at the
top level instead, hence the apparently redundant check for
have_toplevel_build inside the loop here.

Fixes https://github.com/endlessm/amalgamate-pages/issues/14