Fail if GitHub Pages is not correctly configured
Previously, the actions/deploy-pages@v4 step would fail with a
hard-to-understand error message if the repo is not configured to have
GitHub Pages published from Actions.

Add a check at the start of the composite action, and fail the whole job
if the repo is misconfigured, having emitted a summary explaining the
problem. Since Actions workflows are disabled on forks by default, it
seems OK to make this a hard failure: if someone has gone to the trouble
of enabling Actions on their fork of a project using this action, then
they presumably would like to be told why it isn't working.