Include build expiry time as well as date
Previously, I opted to only show the date that an expired build expired.
I think my thinking was to avoid having to think too hard about
timezones!

But instead, show the date and time, in UTC (or more accurately in the
timezone that the GitHub API returns, which is UTC). The same filter
will be used for the timestamp of live builds in future.

You could imagine, some day, using the Power of JavaScript to convert
the datetime (which is available in the <time> element's datetime
attribute in machine-readable form) to a local datetime, or perhaps a
relative time (e.g. "4 days ago") with local time in tooltip.