Link to workflow run & built commit
Showing the commit that the build corresponds to may help to clarify the
linkage between the two, as may the workflow run link.

It is more correct to link to the commit than to the branch because the
branch could have been updated since the site was last regenerated.
(Though you might imagine wanting to link to the branch too.)

In the case of an expired build, linking to the workflow run also
provides a path to triggering a rebuild.