Branches page: Add CSS
Add max width to the body with automatic horizontal margin, so the list
is centered in large screens.

Make the branch name (1st sentence) more important with bigger font
size.

Make hyperlinks for pull request and for builds (2nd and 3rd sentences)
less important so they don't compete with the link to the branch, using
a different style: underline, and turn the color on mouse hover.

Style the list so it doesn't have bullets, and has a bottom border for
separation.

Use different colors for dark mode.