Remove org prefix for branches not on forks
Previously, the link for each playable branch was always prefixed by the
organisation name, e.g. `endlessm/change-repel-animation` or
`wjt/my-beautiful-horse`.

In the common case where the branch is in the repo where the action is
running, rather than in a fork, this doesn't provide any extra
information, and clutters the branch listing.

Only show the organisation when the branch is on a fork.

Helps https://github.com/endlessm/amalgamate-pages/issues/42