Improve error reporting when workflow isn't found
Previously, unless you looked at the output of the script (and
interpreted the Python backtrace), in the case where the wrong workflow
name is given, it has never run, or doesn't exist, the only visible
error was "Process completed with exit code 1."

It is not possible to remove the "Process completed with exit code 1."
message, but it is possible to show an error message as an annotation
next to it. Do so, with a clearer message.

Fixes https://github.com/endlessm/amalgamate-pages/issues/48