Move GitHub Pages configuration check to Python code
Previously, the check was done in a shell script in the action YAML.

I rewrote this check in Python as part of trying to automatically fix
the configuration before realising that workflow jobs don't have enough
permissions to do this.

The rewrite is longer but I think clearer.

Helps https://github.com/endlessm/threadbare/issues/991