Comment on PRs with link to branch's web build
While amalgamating the site, build a list of pull requests and
corresponding web URL, which may be null if the branch is being excluded
due to the PR being closed. Persist this list to a JSON file in the
current directory (which we add to .gitignore).

After the site is published, ingest this list and comment on each PR.
Find an existing issue comment using a magic HTML comment; update the
comment if found, otherwise create a new comment. If an existing comment
is found for a closed PR, update that comment to erase the now-dead
link.

This requires the action to have write permission to pull requests;
update the documentation accordingly.

The JSON file is unversioned because it is not persistent state: it will
always be generated by the same script that consumes it.

Fixes https://github.com/endlessm/amalgamate-pages/issues/18