Factor out GitHubApi class
This just wraps an appropriately-configured Session, along with the
paginate() helper method. It will be used in a future commit which needs
to access a paginated API method outside the AmalgamatePages class.

It may make sense one day to use PyGithub or some other fancier Python
wrapper for the GitHub API, though personally I like the simplicity of
bare API calls.