Factor out iterating over forks × branches
Previously the bulk of the run() function was a doubly-nested loop.
This made it hard to read, in part because the iteration is not trivial,
and partly because the indentation causes moderately-long lines in the
loop body to be wrapped.

Extract an iterator which contains the doubly-nested loop (and the
condition for skipping some branches) and yields a tuple.