Move pull request lookup into iterator
This further simplifies the body of the run() loop.

I did not move the logic that skips closed pull requests into the
iterator because the ultimate goal of these refactorings is to comment
on each pull request with a link to the built branch, and I want to be
able to update the comment when the build is deleted because the PR is
merged.