Build to a fixed directory
Previously, a temporary directory was created each time the script was
run. This ensures that there are no stray files within but is really
annoying for development (because you can't just hit refresh in your
browser, you have to navigate to a different directory) and is not
necessary in CI (because the action runs in a clean checkout).

Instead, build to a `_build` directory next to the script, creating it
if it doesn't exist and deleting and recreating it if it does. Add it to
gitignore.

No change is needed to action.yml: the script still emits the build
directory to $GITHUB_OUTPUT.