Explicitly disable HTTP response cache in CI
The cache expiration logic is not sufficient to avoid the script using
stale responses in some cases. This is good enough for local
development, where the goal of the cache is to reduce the cycle time
when rerunning the script, but it would be bad to accidentally use the
cache in CI.

Happily, because GitHub Actions run in a fresh environment each time,
the cache is ephemeral. But in that case there is no point spending the
electrons to cache the responses which we are going to throw away.

Disable caching if the `CI` environment variable is set to `true`. Use a
plain requests.Session() in that case.

https://docs.github.com/en/actions/reference/workflows-and-actions/variables#default-environment-variables