Set commit statuses as well as commenting on PRs
Setting the commit status, with the test build's playable URL as the
`target_url` for the status, provides a way to find the URL without
having to search a possibly long comment thread in the PR to find the
link. (This is what readthedocs.org does for test builds of PRs.)

This also has the advantage of being something we can set for builds
from branches that do not have an associated PR.

For now, only set the `success` state. You could imagine, in future,
setting `pending` when the site is being rebuilt, but this is
potentially misleading if the branch is already available but the site
is being rebuilt due to a different branch changing.

There is no way to delete a commit status after it is set. We will just
accept broken links in this case, which is the status quo.
(https://github.com/endlessm/amalgamate-pages/issues/16 tracks improving
this.)

This needs a new permission. Rather than bumping the major version, log
a warning and continue if setting a build status fails. Give the same
treatment to PR comments.

Fixes https://github.com/endlessm/amalgamate-pages/issues/63