Set commit status rather than commenting on PRs
Previously, the action would comment on pull requests with a link to the
web build. This is potentially confusing when someone manually comments
with a deep link to a particular scene.

And there is a neater way to report that the build has been published:
setting the commit status, with the test build's playable URL as the
`target_url` for the status. (This is what readthedocs.org does for test
builds of PRs.)

This also has the advantage of being something we can set for builds
from branches that do not have an associated PR.

For now, only set the `success` state. You could imagine, in future,
setting `pending` when the site is being rebuilt, but this is
potentially misleading if the branch is already available but the site
is being rebuilt due to a different branch changing.

There is no way to delete a commit status after it is set. We will just
accept broken links in this case, which is the status quo.
(https://github.com/endlessm/amalgamate-pages/issues/16 tracks improving
this.)

Annoyingly this needs a new permission, so we will once again have to
bump the major version.

Fixes https://github.com/endlessm/amalgamate-pages/issues/63