Ignore PRs with default branch as source
If a branch has no open PRs and 1 or more closed PRs, then it is
excluded from the amalgamated site.

This is generally what you want for topic branches, but if you
accidentally open a PR from your main branch to another branch in your
repo, you do not want your main branch to be hidden forever when you
realise your mistake and close that PR; nor do you want the accidental
PR shown next to it in the branch list forever.

Ignore PRs where the default branch (i.e. main) is the source.

Resolves https://github.com/endlessm/amalgamate-pages/issues/60