Don't set status if existing status is the same
Previously we would unconditionally set a new status for each commit
each time the job runs.

https://docs.github.com/en/rest/commits/statuses?apiVersion=2022-11-28#create-a-commit-status says:

> Note: there is a limit of 1000 statuses per sha and context within a
> repository. Attempts to create more than 1000 statuses will result in
> a validation error.

We appear to be hitting this limit for a very long-lived PR in
Threadbare.

Check the combined status for the commit, which includes the latest
status for each context. Update the status only if there is no existing
status for our context, or if the existing status does not match. (For
instance, the URL may have changed.)

Resolves #77